Add linter from YDB toolkit

+++ cmd/agent/api/urlParams.go
@@ -20,10 +20,10 @@ func (a *api) sendMetrics() {
 			var val string
 			switch m.MType {
 			case "gauge":
-				val = strconv.FormatFloat(float64(*m.Value), 'f', 3, 64)
+				val = strconv.FormatFloat(*m.Value, 'f', 3, 64)
 			case "counter":
 				log.Printf("%v (%v): %+v\n", m.ID, m.MType, m.Value)
-				val = strconv.FormatInt(int64(*m.Delta), 10)
+				val = strconv.FormatInt(*m.Delta, 10)
 			default:
 				log.Printf("Unknown metric type: %#v", m)
 			}
